airline_ml/pyproject.toml
[tool.poetry]
name = "airline-ml"
version = "1.0.7"
description = ""
authors = ["Stef Ruinard <ruinard@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
google-cloud-aiplatform = "^1.67.1"
pyyaml = "^6.0.2"
kfp = "2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
